+++ src/UI/TransactionsTable/components/tests/mockTransactions.ts
@@ -1082,5 +1082,276 @@ export default [
         }
       }
     }
+  },
+  {
+    txHash: '02c3c2660f6c2bcfa767f9539201cd9062ab0f398c3db8bcdca1ae2cd718c639',
+    gasLimit: 75000000,
+    gasPrice: 1000000000,
+    gasUsed: 7953928,
+    miniBlockHash:
+      '8c56f215588dd57815a0ca0147d643b958eda75ddbafb95fa24a84a9c98101e2',
+    nonce: 448,
+    receiver: 'erd1qqqqqqqqqqqqqpgqp699jngundfqw07d8jzkepucvpzush6k3wvqyc44rx',
+    receiverShard: 0,
+    round: 1812254,
+    sender: 'erd1wh9c0sjr2xn8hzf02lwwcr4jk2s84tat9ud2kaq6zr7xzpvl9l5q8awmex',
+    senderShard: 0,
+    signature:
+      '425d8db21db62692983095d9a2148bd0c12a337c743b50a480961ae0f135c1e48a7bcd3bb4e9eb5ce11bd8f96537aefd5660152748d7a09e294e90cb543d7904',
+    status: 'success',
+    value: '0',
+    fee: '167649280000000',
+    timestamp: 1664195124,
+    data: 'dW5TdGFrZUAwMTE1OGU0NjA5MTNkMDAwMDA=',
+    function: 'unStake',
+    action: {
+      category: 'scCall',
+      name: 'unStake'
+    }
+  },
+  {
+    txHash: 'ce1c1d6c4b36a3f239f822307624491cae6f0c4872970a3d0d02c3cc48b2844f',
+    gasLimit: 6000000,
+    gasPrice: 1000000000,
+    gasUsed: 1068000,
+    miniBlockHash:
+      'c72fca9867293054d9182aad167674142cf57c2fc3958282792ecddc55a74cca',
+    nonce: 446,
+    receiver: 'erd1qqqqqqqqqqqqqqqpqqqqqqqqqqqqqqqqqqqqqqqqqqqqqphllllsndz99p',
+    receiverShard: 4294967295,
+    round: 1812226,
+    sender: 'erd1wh9c0sjr2xn8hzf02lwwcr4jk2s84tat9ud2kaq6zr7xzpvl9l5q8awmex',
+    senderShard: 0,
+    signature:
+      '6ed3d00d103e1ade45bd9a6e5ca8194661e414e5a2c8a30cf3360ea9bf6ccc71f14cbc56a68eefbc020502f9a2835278c9cb027c6407face5c6dc225e85c430e',
+    status: 'success',
+    value: '0',
+    fee: '78000000000000',
+    timestamp: 1664194956,
+    data: 'Y2xhaW1SZXdhcmRz',
+    function: 'claimRewards',
+    action: {
+      category: 'stake',
+      name: 'claimRewards',
+      description:
+        'Claim rewards from staking provider https://keybase.io/drfonix',
+      arguments: {
+        providerName: 'https://keybase.io/drfonix'
+      }
+    },
+    results: [
+      {
+        hash:
+          '03ec460f4037114403828f357dc91495149ab35c3556c6ad0155acd8c65c4b04',
+        timestamp: 1664194968,
+        nonce: 0,
+        gasLimit: 0,
+        gasPrice: 1000000000,
+        value: '67861704531764158',
+        sender:
+          'erd1qqqqqqqqqqqqqqqpqqqqqqqqqqqqqqqqqqqqqqqqqqqqqphllllsndz99p',
+        receiver:
+          'erd1wh9c0sjr2xn8hzf02lwwcr4jk2s84tat9ud2kaq6zr7xzpvl9l5q8awmex',
+        prevTxHash:
+          'ce1c1d6c4b36a3f239f822307624491cae6f0c4872970a3d0d02c3cc48b2844f',
+        originalTxHash:
+          'ce1c1d6c4b36a3f239f822307624491cae6f0c4872970a3d0d02c3cc48b2844f',
+        callType: '0',
+        miniBlockHash:
+          '8da210d07a8e099ced579f7ff634b3159f9589cbad106c55c51b4e1c5bf98c7d'
+      },
+      {
+        hash:
+          'ff5d9b11e3fddf3ee4bdf3dec0e82915f1c0b36233e1d05366b9950d78f96fea',
+        timestamp: 1664194968,
+        nonce: 447,
+        gasLimit: 0,
+        gasPrice: 1000000000,
+        value: '49320000000000',
+        sender:
+          'erd1qqqqqqqqqqqqqqqpqqqqqqqqqqqqqqqqqqqqqqqqqqqqqphllllsndz99p',
+        receiver:
+          'erd1wh9c0sjr2xn8hzf02lwwcr4jk2s84tat9ud2kaq6zr7xzpvl9l5q8awmex',
+        data: 'QDZmNmI=',
+        prevTxHash:
+          'ce1c1d6c4b36a3f239f822307624491cae6f0c4872970a3d0d02c3cc48b2844f',
+        originalTxHash:
+          'ce1c1d6c4b36a3f239f822307624491cae6f0c4872970a3d0d02c3cc48b2844f',
+        callType: '0',
+        miniBlockHash:
+          '8da210d07a8e099ced579f7ff634b3159f9589cbad106c55c51b4e1c5bf98c7d'
+      }
+    ],
+    operations: [
+      {
+        id: '03ec460f4037114403828f357dc91495149ab35c3556c6ad0155acd8c65c4b04',
+        action: 'transfer',
+        type: 'egld',
+        sender:
+          'erd1qqqqqqqqqqqqqqqpqqqqqqqqqqqqqqqqqqqqqqqqqqqqqphllllsndz99p',
+        receiver:
+          'erd1wh9c0sjr2xn8hzf02lwwcr4jk2s84tat9ud2kaq6zr7xzpvl9l5q8awmex',
+        value: '67861704531764158'
+      }
+    ]
+  },
+  {
+    txHash: '5e73704d40fdef680bd8839cd277827209f5decd3679b235efddd921f1c9fb3b',
+    gasLimit: 12000000,
+    gasPrice: 1000000000,
+    gasUsed: 6075500,
+    miniBlockHash:
+      '3a7470a369d5e5bf6b5e984df1f6fdbed2fb260bb81574b1392e8f42d3c17f96',
+    nonce: 445,
+    receiver: 'erd1qqqqqqqqqqqqqqqpqqqqqqqqqqqqqqqqqqqqqqqqqqqqq9lllllsf3mp40',
+    receiverShard: 4294967295,
+    round: 1809456,
+    sender: 'erd1wh9c0sjr2xn8hzf02lwwcr4jk2s84tat9ud2kaq6zr7xzpvl9l5q8awmex',
+    senderShard: 0,
+    signature:
+      '3a25067722f2c8a98f693ccb2861f6ea1f91565d99e6b6a1aee57788d612a78d22cf1c3bfb55e94d9f5e881efad8632a4ad5be83e0830ad8f6181c8ac1552205',
+    status: 'success',
+    value: '0',
+    fee: '135500000000000',
+    timestamp: 1664178336,
+    data: 'cmVEZWxlZ2F0ZVJld2FyZHM=',
+    function: 'reDelegateRewards',
+    action: {
+      category: 'stake',
+      name: 'reDelegateRewards',
+      description:
+        'Redelegate rewards from staking provider figment.io_testnet',
+      arguments: {
+        providerName: 'figment.io_testnet',
+        providerAvatar:
+          'https://s3.amazonaws.com/keybase_processed_uploads/8cba5dc15c79565402849fd869654705_360_360.jpg'
+      }
+    },
+    results: [
+      {
+        hash:
+          '1bf45b782580cd035efdf600aafbfdd6d3fc437e93b4fb9219fe76f7526fad65',
+        timestamp: 1664178336,
+        nonce: 0,
+        gasLimit: 0,
+        gasPrice: 1000000000,
+        value: '392502810808819111',
+        sender:
+          'erd1qqqqqqqqqqqqqqqpqqqqqqqqqqqqqqqqqqqqqqqqqqqqq9lllllsf3mp40',
+        receiver:
+          'erd1qqqqqqqqqqqqqqqpqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqplllst77y4l',
+        prevTxHash:
+          '5e73704d40fdef680bd8839cd277827209f5decd3679b235efddd921f1c9fb3b',
+        originalTxHash:
+          '5e73704d40fdef680bd8839cd277827209f5decd3679b235efddd921f1c9fb3b',
+        callType: '0',
+        miniBlockHash:
+          'e74217b57b02069cf1b3294db1acbce08c5174811464c283cd1c739dbbf27b74'
+      },
+      {
+        hash:
+          '5fef839b3abc2f07be0923acfcb086a19ad2cf9522e9cf2eee4f398c84b301ef',
+        timestamp: 1664178348,
+        nonce: 446,
+        gasLimit: 0,
+        gasPrice: 1000000000,
+        value: '59245000000000',
+        sender:
+          'erd1qqqqqqqqqqqqqqqpqqqqqqqqqqqqqqqqqqqqqqqqqqqqq9lllllsf3mp40',
+        receiver:
+          'erd1wh9c0sjr2xn8hzf02lwwcr4jk2s84tat9ud2kaq6zr7xzpvl9l5q8awmex',
+        data: 'QDZmNmI=',
+        prevTxHash:
+          '5e73704d40fdef680bd8839cd277827209f5decd3679b235efddd921f1c9fb3b',
+        originalTxHash:
+          '5e73704d40fdef680bd8839cd277827209f5decd3679b235efddd921f1c9fb3b',
+        callType: '0',
+        miniBlockHash:
+          '8f15889818ae5f2b63bebfae42deb1265ef500832aa5d12510246535140e1465'
+      }
+    ],
+    operations: [
+      {
+        id: '1bf45b782580cd035efdf600aafbfdd6d3fc437e93b4fb9219fe76f7526fad65',
+        action: 'transfer',
+        type: 'egld',
+        sender:
+          'erd1qqqqqqqqqqqqqqqpqqqqqqqqqqqqqqqqqqqqqqqqqqqqq9lllllsf3mp40',
+        receiver:
+          'erd1qqqqqqqqqqqqqqqpqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqplllst77y4l',
+        value: '392502810808819111'
+      }
+    ]
+  },
+  {
+    txHash: '5bdfd566dcc2c19a8bd62caa138acd2662a9d5f3c76213698e11c0f74c28cda6',
+    gasLimit: 75000000,
+    gasPrice: 1000000000,
+    gasUsed: 75000000,
+    miniBlockHash:
+      'bde6aad6833c6eac6a58f26fbed6ae3960615b900ba41cf634138eb3c4902a1e',
+    nonce: 449,
+    receiver: 'erd1qqqqqqqqqqqqqpgqp699jngundfqw07d8jzkepucvpzush6k3wvqyc44rx',
+    receiverShard: 0,
+    round: 1813482,
+    sender: 'erd1wh9c0sjr2xn8hzf02lwwcr4jk2s84tat9ud2kaq6zr7xzpvl9l5q8awmex',
+    senderShard: 0,
+    signature:
+      '414b703c1e998daacf0d7ddf809b88ccd9ded2e673dc4ff978ec1c01f59a2457ddef21196ae7c566e2268be4fd0cb09be876cca938352f7acd6793b377c99e0f',
+    status: 'success',
+    value: '0',
+    fee: '808410000000000',
+    timestamp: 1664202492,
+    data: 'dW5Cb25k',
+    function: 'unBond',
+    action: {
+      category: 'scCall',
+      name: 'unBond'
+    },
+    results: [
+      {
+        hash:
+          '62ad6e43a172cfc39ae1620e61a0e200cb5cd51e9cf1675a19cc4963971fdc65',
+        timestamp: 1664202492,
+        nonce: 0,
+        gasLimit: 0,
+        gasPrice: 1000000000,
+        value: '20000000000000000000',
+        sender:
+          'erd1qqqqqqqqqqqqqpgqp699jngundfqw07d8jzkepucvpzush6k3wvqyc44rx',
+        receiver:
+          'erd1wh9c0sjr2xn8hzf02lwwcr4jk2s84tat9ud2kaq6zr7xzpvl9l5q8awmex',
+        data: 'ZGVsZWdhdGlvbiBzdGFrZSB1bmJvbmQ=',
+        prevTxHash:
+          '5bdfd566dcc2c19a8bd62caa138acd2662a9d5f3c76213698e11c0f74c28cda6',
+        originalTxHash:
+          '5bdfd566dcc2c19a8bd62caa138acd2662a9d5f3c76213698e11c0f74c28cda6',
+        callType: '0'
+      }
+    ],
+    operations: [
+      {
+        id: '5bdfd566dcc2c19a8bd62caa138acd2662a9d5f3c76213698e11c0f74c28cda6',
+        action: 'transfer',
+        type: 'egld',
+        sender:
+          'erd1qqqqqqqqqqqqqpgqp699jngundfqw07d8jzkepucvpzush6k3wvqyc44rx',
+        receiver:
+          'erd1wh9c0sjr2xn8hzf02lwwcr4jk2s84tat9ud2kaq6zr7xzpvl9l5q8awmex',
+        value: '20000000000000000000'
+      },
+      {
+        id: '5bdfd566dcc2c19a8bd62caa138acd2662a9d5f3c76213698e11c0f74c28cda6',
+        action: 'writeLog',
+        type: 'log',
+        sender:
+          'erd1qqqqqqqqqqqqqpgqp699jngundfqw07d8jzkepucvpzush6k3wvqyc44rx',
+        receiver:
+          'erd1wh9c0sjr2xn8hzf02lwwcr4jk2s84tat9ud2kaq6zr7xzpvl9l5q8awmex',
+        data: '@6f6b@01158e460913d00000',
+        message:
+          '@too much gas provided for processing: gas provided = 74941000, gas used = 6578498'
+      }
+    ]
   }
 ];